Rearrange server starting and including so that importing "stripe-stateful-mock" doesn't start the server automatically because that's counter-intuitive.  Add the necessary magic to be run from the command line. Update README to reflect new usage.
