+++ src/cli.ts
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+// Installed by npm as a script that can be run from the command line.
+// See https://docs.npmjs.com/files/package.json#bin
+
+require("./autostart");
